data-migration-py/datamigration-csv-tochain/covey/requirements.txt
skale.py==4.1b22
python-dotenv==0.14.0
requests~=2.23.0
web3~=5.8.0
